Reworked the navigation route
Did some tinkering around and made a less convoluted navigation route for the app
