--- app/screens/LoginScreen.tsx
@@ -1,40 +1,0 @@
-import React from 'react';
-import { View, Button, StyleSheet } from 'react-native';
-import * as Google from 'expo-auth-session/providers/google';
-import Constants from 'expo-constants';
-
-export default function LoginScreen() {
-  const [request, response, promptAsync] = Google.useAuthRequest({
-    clientId: Constants.expoConfig?.extra?.expoClientId,
-    iosClientId: Constants.expoConfig?.extra?.iosClientId,
-    androidClientId: Constants.expoConfig?.extra?.androidClientId,
-    webClientId: Constants.expoConfig?.extra?.webClientId,
-  });
-
-  React.useEffect(() => {
-    if (response?.type === 'success') {
-      const { authentication } = response;
-      // Handle successful authentication here
-    }
-  }, [response]);
-
-  return (
-    <View style={styles.container}>
-      <Button
-        disabled={!request}
-        title="Login with Google"
-        onPress={() => {
-          promptAsync();
-        }}
-      />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
